Move VelocyPackStringLiteral into new velocypack_utils library
