+++ lib/VelocypackUtils/VelocyPackStringLiteral.cpp
